+++ packages/learn-card-core/src/wallet/plugins/vc/issuePresentation.ts
@@ -1,37 +1,32 @@
-import { issuePresentation as ip, keyToVerificationMethod } from 'didkit';
-import { VC } from '@learncard/types';
+import { VC, UnsignedVP } from '@learncard/types';
 
+import { DependentMethods, VCPluginMethods } from './types';
 import { UnlockedWallet } from 'types/wallet';
 
-export const issuePresentation = async (
-    wallet: UnlockedWallet<
-        any,
-        { getSubjectDid: () => string; getSubjectKeypair: () => Record<string, string> }
-    >,
-    credential: VC
-) => {
-    const did = wallet.pluginMethods.getSubjectDid();
+export const issuePresentation = (initWallet: UnlockedWallet<string, DependentMethods>) => {
+    return async (wallet: UnlockedWallet<string, VCPluginMethods>, credential: VC) => {
+        const did = wallet.pluginMethods.getSubjectDid('key');
 
-    if (!did) throw new Error('Cannot create presentation: No holder key found');
+        if (!did) throw new Error('Cannot create presentation: No holder key found');
 
-    const holder = did;
+        const holder = did;
 
-    const _kp = wallet.pluginMethods.getSubjectKeypair();
+        const kp = wallet.pluginMethods.getSubjectKeypair();
 
-    if (!_kp) throw new Error('Cannot issue credential: Could not get subject keypair');
+        if (!kp) throw new Error('Cannot issue credential: Could not get subject keypair');
 
-    const kp = JSON.stringify(_kp);
-    const options = JSON.stringify({
-        verificationMethod: await keyToVerificationMethod('key', kp),
-        proofPurpose: 'assertionMethod',
-    });
+        const options = {
+            verificationMethod: await initWallet.pluginMethods.keyToVerificationMethod('key', kp),
+            proofPurpose: 'assertionMethod',
+        };
 
-    const presentation = JSON.stringify({
-        '@context': ['https://www.w3.org/2018/credentials/v1'],
-        type: ['VerifiablePresentation'],
-        holder,
-        verifiableCredential: credential,
-    });
+        const presentation: UnsignedVP = {
+            '@context': ['https://www.w3.org/2018/credentials/v1'],
+            type: ['VerifiablePresentation'],
+            holder,
+            verifiableCredential: credential,
+        };
 
-    return JSON.parse(await ip(presentation, options, kp));
+        return initWallet.pluginMethods.issuePresentation(presentation, options, kp);
+    };
 };