+++ packages/learn-card-core/src/wallet/plugins/vc/types.ts
@@ -1,21 +1,30 @@
-import { UnsignedVC, VC, VP } from '@learncard/types';
+import { UnsignedVC, VC, UnsignedVP, VP, VerificationCheck } from '@learncard/types';
+import { KeyPair, ProofOptions } from '../didkit/types';
 
-export type VerificationCheck = {
-    checks: string[];
-    warnings: string[];
-    errors: string[];
+export type DependentMethods = {
+    getSubjectDid: (type: 'key') => string;
+    getSubjectKeypair: () => KeyPair;
+    keyToVerificationMethod: (type: string, keypair: KeyPair) => Promise<string>;
+    issueCredential: (
+        credential: UnsignedVC,
+        options: ProofOptions,
+        keypair: KeyPair
+    ) => Promise<VC>;
+    verifyCredential: (credential: VC) => Promise<VerificationCheck>;
+    issuePresentation: (
+        presentation: UnsignedVP,
+        options: ProofOptions,
+        keypair: KeyPair
+    ) => Promise<VP>;
+    verifyPresentation: (presentation: VP) => Promise<VerificationCheck>;
 };
 
-export type VCPluginMethods = {
+export type VCPluginMethods = DependentMethods & {
     issueCredential: (credential: UnsignedVC) => Promise<VC>;
     verifyCredential: (credential: VC) => Promise<VerificationCheck>;
     issuePresentation: (credential: VC) => Promise<VP>;
     verifyPresentation: (presentation: VP) => Promise<VerificationCheck>;
     getTestVc: (subject?: string) => UnsignedVC;
-
-    // Dependent methods
-    getSubjectDid: () => string;
-    getSubjectKeypair: () => Record<string, string>;
 };
 
 export type VerifyExtension = { verifyCredential: (credential: VC) => Promise<VerificationCheck> };